refactor : 일부 참조 및 선언 최적화, /calendar 페이지 삭제 및 /로 통일
