+++ components/event-dialog.tsx
@@ -6,7 +6,7 @@ import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { Textarea } from "@/components/ui/textarea"
-import type { Event } from "@/app/page"
+import type { Event } from "@/types/calendar"
 
 type EventDialogProps = {
   open: boolean
